feat(): new chat stream
